chore: move logic to contact form page
